Cache last-sent response
This fixes the "survey" box not leaving bug. Generally, when a button is pressed, streamlit reruns the entire state of the app at once. This causes the query the user inputted before replying to the survey be re-generated.

These changes cache the most recently sent query. So, when streamlit reruns the entire code and it reaches the `if query`, it wont actually have a new query. This is because it initially checks for a boolean "is_new_query". If it isnt a new query, the if statement will not rerun and regenerate the same response redundantly.

Also, the previous commit that removed the survey box did not work since it tracked the actual survey response which wasnt initialized. Now it checks for a db_id which always exists for new survey instances.